rpctest/requirements.txt
pycurl
python-bitcoinrpc>=1.0
python-dateutil>=2.6.1
PyYAML>=3.12
requests>=2.18.4
requests-unixsocket>=0.1.5
slick-bitcoinrpc>=0.1.4
ujson>=1.35
urllib3>=1.22
pycrypto>=2.6.1
pytest>=5.0.1
python-daemon
